put script and requirements into subfolder
