--- tools/requirements.txt
@@ -1,3 +1,0 @@
-matplotlib
-mpld3
-tornado
