+++ tools/wt_verify/requirements.txt
@@ -1,0 +1,3 @@
+matplotlib
+mpld3
+tornado
